+++ update_warehouse_lflive.py
@@ -1,5 +1,5 @@
 
-def update_warehouse_lflive(table_name, temporary_table_name, wh_query, wh_combined_table,
+def update_warehouse_lflive(table_name, temporary_table_name, wh_query, wh_combined_table, delete_staging,
 	print_internal=False, print_details=False):
 	
 	if print_details == True:
@@ -107,10 +107,10 @@ def update_warehouse_lflive(table_name, temporary_table_name, wh_query, wh_combi
 	query_database2('drop External Table '+temporary_table_name,drop_sql, 
 		output_db, output_database, print_details=print_details, ignore_errors=True)	
 
-	drop_sql = "DROP TABLE "+wh_combined_table
-	print(drop_sql)
-	query_database2('drop Table '+wh_combined_table,drop_sql, 
-		output_db, output_database, print_details=print_details, ignore_errors=True)
+	if delete_staging == True:
+		drop_sql = "DROP TABLE "+wh_combined_table
+		query_database2('drop Table '+wh_combined_table,drop_sql, 
+			output_db, output_database, print_details=print_details, ignore_errors=True)
 	
 	""""""
